requirements.txt
alembic==0.9.9
blinker==1.4
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
click==6.7
cssselect2==0.2.2
defusedxml==0.6.0
Flask==1.0.2
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WeasyPrint==0.6
Flask-WTF==0.14
gevent==1.4.0
greenlet==0.4.15
html2pdf==0.1
html5lib==1.0.1
idna==2.8
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
mimerender==0.6.0
mysql-connector==2.2.9
numpy==1.17.4
pandas==0.25.3
pdfkit==0.6.1
Pillow==5.3.0
pkg-resources==0.0.0
py-edamam==0.2
pycparser==2.19
PyMySQL==0.8.1
Pyphen==0.9.5
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
python-mimeparse==1.6.0
python-slugify==1.2.5
pytz==2019.3
requests==2.22.0
simplejson==3.16.0
six==1.11.0
SQLAlchemy==1.2.8
tinycss2==1.0.2
Unidecode==1.0.22
urllib3==1.25.6
WeasyPrint==51
webencodings==0.5.1
Werkzeug==0.14.1
wkhtmltopdf==0.2
WTForms==2.2.1
xlrd==1.2.0
